cmt 28/3

+++ src/main/webapp/template/admin/js/building-list.js
@@ -1,7 +1,8 @@
 function assingmentBuilding(buildingId){
 	openModelAssingmentBuilding();		
 	$('#buildingId').val(buildingId);
-	console.log($('#buildingId').val());		
+	console.log($('#buildingId').val());	
+	showStaffAssignment();
 }
 function openModelAssingmentBuilding() { 
 	$('#assingmentBuildingModel').modal();
@@ -34,14 +35,45 @@ function assignStaff(data){
 		}
 	});
 }
+function showStaffAssignment(){
+	$.ajax({
+		type: "GET",
+		url: "http://localhost:8080/api-user?type=SHOW_STAFF_ASSIGNMENT",		
+		dataType: "json",
+		success: function (response) {
+			console.log('success!');
+			console.log(response);
+			var html = '';
+			$.each(response, function (index, staffOutput) { 
+				html += '<tr>';
+				html += '<td><input type="checkbox" value="'+staffOutput.staffId+'" id="checkbox_'+staffOutput.staffId+'" '+staffOutput.checked+'></td>';
+				html += '<td>'+staffOutput.fullName+'</td>';
+				html += '</tr>';
+			});			
+			$('#staffList tbody').html(html);
+		},
+		error: function (response) {
+			console.log('failed!');
+			console.log(response);
+		}
+	});
+}
 $('#btnDeleteBuilding').click(function (e) { 
 	e.preventDefault();
+	if (!confirm("Bạn chắc chắn muốn xóa không?")) {
+	    return;
+	}
 	var data = {};			
 	//$('#staffList').find('tbody input[type=checkbox]');
 	var buildingIds = $('#buildingList').find('tbody input[type=checkbox]:checked').map(function () {
 		return $(this).val();
 	}).get();
-	data['buildingIds'] =  buildingIds;
+	data['ids'] =  buildingIds;
+	if(buildingIds.length == 0){
+		//showModal("", "Chưa có tòa nhà nào được chọn.")
+		alert("Chưa có tòa nhà nào được chọn.")
+		return;
+	}
 	deleteBuilding(data);
 });
 function deleteBuilding(data){
@@ -49,13 +81,16 @@ function deleteBuilding(data){
 		type: "DELETE",
 		url: "http://localhost:8080/api-building",
 		data: JSON.stringify(data),				
-		dataType: "application/json",
+		dataType: "json",
 		success: function (response) {
 			console.log('success!');
-			console.log(response);
+			console.log(response);	
+			alert("Xóa thành công!");
+			location.reload();
 		},
 		error: function (response) {
 			console.log('failed!');
+			alert("Đã có lỗi xảy ra!" + response);
 			console.log(response);
 		}
 	});
@@ -64,3 +99,12 @@ $('#btnSearchBuilding').click(function (e) {
 	e.preventDefault();
 	$('#formSearchBuilding').submit();
 });
+
+function removeBuilding(buildingId){
+	if (!confirm("Bạn chắc chắn muốn xóa không?")) {
+	    return;
+	}
+	var data = {};	
+	data['ids'] =  [buildingId];
+	deleteBuilding(data);
+}
